Refactor from pandas tests
